Added Version 2.2
Contains keyboard support: User can press any of the appropriate buttons
and the program will validate and display it to the screen.
I also included an information button that gives the user a bit more
details about which keys are supported. The button can be found to the
right of the calculator.

While implementing the above, I decided to modify the setDisplay()
function rather than create a new one that can support keys.
To aid this change, I created validateKey() which verifys that a key
is legal before passing it to setDisplay(). I felt that this
path would be more efficient as each function is now limited to a
single task."